Refactored tests down to 1 file.
