--- packages/components/organisms/f-checkout/test/specs/component/f-checkout-guest-delivery.component.spec.js
@@ -1,72 +1,0 @@
-import forEach from 'mocha-each';
-import CheckoutComponent from '../../../test-utils/component-objects/f-checkout.component';
-
-describe('f-checkout "delivery" component tests', () => {
-    before(() => {
-        browser.url('iframe.html?id=components-organisms--checkout-component&knob-Checkout%20Url=%2Fcheckout-delivery.json&knob-Available%20Fulfilment%20Url=%2Fcheckout-available-fulfilment.json&knob-Create%20Guest%20Url=%2Fcreate-guest.json&knob-Auth%20token=&knob-Login%20Url=%2Flogin&viewMode=story');
-        CheckoutComponent.waitForCheckoutComponent();
-    });
-
-    forEach(['firstName', 'lastName', 'emailAddress', 'mobileNumber', 'addressLine1', 'addressCity', 'addressPostcode'])
-    .it('should display each fields error message', field => {
-        // Act
-        CheckoutComponent.clearCheckoutForm(field);
-        CheckoutComponent.goToPayment();
-
-        // Assert
-        expect(CheckoutComponent.isFieldErrorDisplayed(field)).toBe(true);
-    });
-
-    forEach(['firstName', 'lastName', 'emailAddress', 'addressLine1', 'addressLine2', 'addressCity', 'addressPostcode'])
-    .it('should show the address fields', field => {
-        // Assert
-        expect(CheckoutComponent.doesFieldExist(field)).toBe(true);
-    });
-
-    it('should display the mandatory fields', () => {
-        // Assert
-        expect(CheckoutComponent.isFieldDisplayed('mobileNumber')).toBe(true);
-    });
-
-    it('should prevent user from submitting an invalid email address', () => {
-        // Arrange
-        const emailAddress = {
-            emailAddress: 'abc@abc'
-        };
-
-        // Act
-        CheckoutComponent.populateCheckoutForm(emailAddress);
-        CheckoutComponent.goToPayment();
-
-        // Assert
-        expect(CheckoutComponent.isFieldErrorDisplayed('emailAddress')).toBe(true);
-    });
-
-    it('should prevent user from submitting a postcode with illegal characters', () => {
-        // Arrange
-        const addressInfo = {
-            postcode: 'TEST1A'
-        };
-
-        // Act
-        CheckoutComponent.populateCheckoutForm(addressInfo);
-        CheckoutComponent.goToPayment();
-
-        // Assert
-        expect(CheckoutComponent.isPostCodeTypeErrorDisplayed()).toBe(true);
-    });
-
-    it('should enable a user to submit a postcode with correct characters', () => {
-        // Arrange
-        const addressInfo = {
-            postcode: 'AR51 1AA'
-        };
-
-        // Act
-        CheckoutComponent.populateCheckoutForm(addressInfo);
-        CheckoutComponent.goToPayment();
-
-        // Assert
-        expect(CheckoutComponent.isPostCodeTypeErrorDisplayed()).toBe(false);
-    });
-});
